Removing warning discovered by LLVM.
git-svn-id: http://svn.seqan.de/seqan/trunk@9481 e6417c60-b987-48fd-844e-b20f0fcc1017

+++ seqan/core/include/seqan/find/find_abndm.h
@@ -282,7 +282,7 @@ inline bool _findAbndmSmallNeedle(TFinder & finder, Pattern<TNeedle, AbndmAlgo> 
         finder -= offset;
         int end = me.cP + me.needleLength + me.limit;
         // adjust end if it points over the edges of host(finder)
-        end = (end > static_cast<int>(length(host(finder))) ? length(host(finder)) : end);
+        end = (end > static_cast<int>(length(host(finder))) ? static_cast<int>(length(host(finder))) : end);
 
         THostSegment s(infix(host(finder),me.cP, end));
         THSFinder f(s);
